+++ validation/src/main/kotlin/de/otto/babbage/validation/SafeId.kt
@@ -1,0 +1,22 @@
+package de.otto.babbage.validation
+
+import jakarta.validation.Constraint
+import jakarta.validation.Payload
+import kotlin.reflect.KClass
+
+@MustBeDocumented
+@Constraint(validatedBy = [SafeIdValidator::class])
+@Target(
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.TYPE
+)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class SafeId(
+    val message: String = "{invalid.id.value}",
+    val groups: Array<KClass<*>> = [],
+    val payload: Array<KClass<out Payload>> = []
+)