+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   * Fix: Logger overview now uses the correct links for changing log levels.
   * Fix: Long loggers are now scrollable and are not hidden behind the log-level buttons.
   * Feature: Add option to add custom management pages by extending your Controller with `ManagementController`. Documentation about is added [here](core/README.md)
+* _validation_: 
+  * add `SafeId` annotation to validation String input parameters 
 
 ## 0.0.2
 * _aws-paramstore:_ Fix: Add missing `@ConfigurationProperties` for AWS ParamStore Properties.